Implements code-reorg
Addresses some PR comments; adds a default warning level log for un-mapped activity types.
