+++ test/support/factory.ex
@@ -139,6 +139,7 @@ defmodule Trento.Factory do
 
   alias Trento.ActivityLog.ActivityLog, as: ActivityLogEntry
   alias Trento.ActivityLog.RetentionTime
+  alias Trento.ActivityLog.SeverityLevel
 
   alias Trento.Abilities.{
     Ability,
@@ -1162,7 +1163,7 @@ defmodule Trento.Factory do
     %ActivityLogEntry{
       type: Faker.Pokemon.name(),
       actor: Faker.Internet.user_name(),
-      severity: ActivityLogEntry.severity_level_to_integer(:info),
+      severity: SeverityLevel.severity_level_to_integer(:info),
       metadata: %{}
     }
   end
